python/src/app/scanner_client.py
"""
gRPC client for the Go scanner service.
"""

import os
from typing import Any, Dict, List, Optional

import grpc
from src.utils.logger import log_debug, log_error, log_info

# Import the generated Python gRPC code (placeholder, would be generated from proto files)
# from ..proto import scanner_pb2, scanner_pb2_grpc


class ScannerClient:
    """gRPC client for the Go scanner service."""

    host: str
    port: int
    channel: Optional[Any]
    stub: Any

    def __init__(self, host: str, port: int) -> None:
        """
        Initialize the scanner client.

        Args:
            host: Scanner service host
            port: Scanner service port
        """
        self.host = host
        self.port = port
        self.channel = None
        self.stub = None

        # Connect to the service
        self.connect()

    def connect(self) -> bool:
        """
        Connect to the scanner service.

        Returns:
            True if connected successfully, False otherwise
        """
        try:
            address = f"{self.host}:{self.port}"
            log_info(f"Connecting to scanner service at {address}")

            # Create gRPC channel
            self.channel = grpc.insecure_channel(address)

            # Create stub (client)
            # self.stub = scanner_pb2_grpc.ScannerServiceStub(self.channel)

            # For now, we'll just simulate this since we don't have the generated code
            self.stub = DummyScannerStub()

            log_info("Connected to scanner service")
            return True
        except Exception as e:
            log_error(f"Failed to connect to scanner service: {str(e)}")
            return False

    def close(self) -> None:
        """Close the connection to the scanner service."""
        if self.channel:
            self.channel.close()
            self.channel = None
            self.stub = None
            log_info("Closed connection to scanner service")

    def scan(self, symbols: List[str], strategies: List[str]) -> Dict[str, List[str]]:
        """
        Scan symbols for trade signals.

        Args:
            symbols: List of symbols to scan
            strategies: List of strategies to apply

        Returns:
            Dictionary mapping symbols to signal types
        """
        try:
            log_debug(
                f"Scanning {len(symbols)} symbols with {len(strategies)} strategies"
            )

            # Create request object
            # request = scanner_pb2.ScanRequest(
            #     symbols=symbols,
            #     strategies=strategies,
            #     date_range=scanner_pb2.DateRange(
            #         start_date="",  # Use empty for latest data
            #         end_date=""
            #     )
            # )

            # Make the gRPC call
            # response = self.stub.Scan(request)

            # For now, simulate a response
            response = self.stub.Scan(symbols, strategies)

            # Process and return results
            # signals = {symbol: list(signal_list.signal_types) for symbol, signal_list in response.signals.items()}
            signals: Dict[str, List[str]] = response

            log_debug(f"Scan completed, found signals for {len(signals)} symbols")
            return signals
        except Exception as e:
            log_error(f"Error scanning symbols: {str(e)}")
            return {}

    def bulk_fetch(self, symbols: List[str], timeframe: str) -> Dict[str, bytes]:
        """
        Fetch historical data for multiple symbols.

        Args:
            symbols: List of symbols to fetch
            timeframe: Data timeframe ("daily", "minute")

        Returns:
            Dictionary mapping symbols to serialized data
        """
        try:
            log_debug(
                f"Bulk fetching {len(symbols)} symbols with timeframe {timeframe}"
            )

            # Create request object
            # request = scanner_pb2.BulkFetchRequest(
            #     symbols=symbols,
            #     timeframe=timeframe,
            #     date_range=scanner_pb2.DateRange(
            #         start_date="",  # Use empty for latest data
            #         end_date=""
            #     )
            # )

            # Make the gRPC call
            # response = self.stub.BulkFetch(request)

            # For now, simulate a response
            response = self.stub.BulkFetch(symbols, timeframe)

            # Process and return results
            # data = dict(response.data)
            data: Dict[str, bytes] = response

            log_debug(f"Bulk fetch completed, got data for {len(data)} symbols")
            return data
        except Exception as e:
            log_error(f"Error fetching data: {str(e)}")
            return {}

    def get_metrics(self) -> Dict[str, float]:
        """
        Get performance metrics from the scanner service.

        Returns:
            Dictionary of metrics
        """
        try:
            # Create request object
            # request = scanner_pb2.MetricsRequest()

            # Make the gRPC call
            # response = self.stub.GetMetrics(request)

            # For now, simulate a response
            response = self.stub.GetMetrics()

            # Process and return results
            metrics = {
                # "avg_scan_time_seconds": response.avg_scan_time_seconds,
                # "symbols_per_second": response.symbols_per_second,
                # "total_scans": response.total_scans,
                # "memory_usage_mb": response.memory_usage_mb,
                # "cpu_usage_percent": response.cpu_usage_percent
                "avg_scan_time_seconds": response.get("avg_scan_time_seconds", 0),
                "symbols_per_second": response.get("symbols_per_second", 0),
                "total_scans": response.get("total_scans", 0),
                "memory_usage_mb": response.get("memory_usage_mb", 0),
                "cpu_usage_percent": response.get("cpu_usage_percent", 0),
            }

            log_debug(f"Got metrics: {metrics}")
            return metrics
        except Exception as e:
            log_error(f"Error getting metrics: {str(e)}")
            return {
                "avg_scan_time_seconds": 0,
                "symbols_per_second": 0,
                "total_scans": 0,
                "memory_usage_mb": 0,
                "cpu_usage_percent": 0,
            }


# Dummy stub for development until we have generated gRPC code
class DummyScannerStub:
    """Dummy stub for development."""

    def Scan(self, symbols: List[str], strategies: List[str]) -> Dict[str, List[str]]:
        """Simulate a scan response."""
        # Return dummy signals for the first few symbols
        signals = {}
        for symbol in symbols[: min(3, len(symbols))]:
            signals[symbol] = ["LONG"] if symbol.startswith("A") else ["SHORT"]
        return signals

    def BulkFetch(self, symbols: List[str], timeframe: str) -> Dict[str, bytes]:
        """Simulate a bulk fetch response."""
        # Return dummy data for each symbol
        data = {}
        for symbol in symbols:
            data[symbol] = b"dummy data"
        return data

    def GetMetrics(self) -> Dict[str, float]:
        """Simulate a metrics response."""
        return {
            "avg_scan_time_seconds": 0.5,
            "symbols_per_second": 100.0,
            "total_scans": 10,
            "memory_usage_mb": 50.0,
            "cpu_usage_percent": 5.0,
        }

