python/pyproject.toml
[tool.poetry]
name = "ibkr-trader"
version = "0.1.0"
description = "Interactive Brokers Trading System"
authors = ["Dan <user@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
pyyaml = "^6.0"
requests = "^2.28.1"
grpcio = "^1.58.0"
grpcio-tools = "^1.58.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
pytz = "^2023.3"
psutil = "^5.9.0"
ib-insync = "^0.9.70"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-p no:mypy"

